v-6-core

+++ Services/FileService.cs
@@ -3,8 +3,8 @@ using MongoDB.Bson;
 using MongoDB.Driver.GridFS;
 using System;
 using System.Collections.Generic;
+using System.IO;
 using System.Linq;
-using System.Text;
 using System.Threading.Tasks;
 
 namespace AmazingFileVersionControl.Core.Services
@@ -19,20 +19,39 @@ namespace AmazingFileVersionControl.Core.Services
         }
 
         public async Task<ObjectId> UploadFileAsync(string name,
-            string owner, string project,
-            string type, Stream stream,
-            string? description = null)
+            string owner, string type,
+            string project, Stream stream,
+            string? description = null, long? version = null)
         {
+            if (version.HasValue && (version.Value < -1 || version.Value == 0))
+            {
+                throw new ArgumentException("Version cannot be less than -1 or equal to 0.");
+            }
+
             try
             {
-                long lastVersionNumber = await GetLastVersionNumberAsync(name, owner, project) + 1;
+                long versionToUse;
+                if (version.HasValue && version.Value != -1)
+                {
+                    var existingFile = await GetFileInfoByVersionAsync(name, owner, type, project, version.Value);
+                    if (existingFile != null)
+                    {
+                        await DeleteFileByVersionAsync(name, owner, type, project, version.Value);
+                    }
+                    versionToUse = version.Value;
+                }
+                else
+                {
+                    long lastVersionNumber = await GetLastVersionNumberAsync(name, owner, type, project);
+                    versionToUse = lastVersionNumber + 1;
+                }
 
                 var metadata = new BsonDocument
                 {
-                    { "file_type", type},
+                    { "file_type", type },
                     { "owner", owner },
                     { "project", project },
-                    { "version_number", lastVersionNumber},
+                    { "version_number", versionToUse },
                     { "version_description", description ?? "No description provided" }
                 };
 
@@ -44,55 +63,68 @@ namespace AmazingFileVersionControl.Core.Services
             }
         }
 
-        public async Task<Stream> DownloadFileAsync(string name,
-            string owner, string project,
-            long version = -1)
+        public async Task<(Stream, BsonDocument)> DownloadFileWithMetadataAsync(string name, string owner, string type, string project, long? version = null)
         {
+            if (version.HasValue && (version.Value < -1 || version.Value == 0))
+            {
+                throw new ArgumentException("Version cannot be less than -1 or equal to 0.");
+            }
+
             try
             {
-                long lastVersionNumber = await GetLastVersionNumberAsync(name, owner, project);
+                long versionToUse = version.HasValue && version.Value != -1
+                                    ? version.Value
+                                    : await GetLastVersionNumberAsync(name, owner, type, project);
 
                 var query = new BsonDocument
                 {
                     { "filename", name },
                     { "metadata.owner", owner },
-                    { "metadata.project", project }
+                    { "metadata.project", project },
+                    { "metadata.file_type", type },
+                    { "metadata.version_number", versionToUse }
                 };
 
-                if (version < 0)
+                var stream = await _fileRepository.DownloadAsync(query);
+                var fileInfo = await _fileRepository.GetOneAsync(query);
+
+                if (fileInfo == null)
                 {
-                    query.Add("metadata.version_number", lastVersionNumber);
-                }
-                else
-                {
-                    query.Add("metadata.version_number", version);
+                    throw new FileNotFoundException("File not found.");
                 }
 
-                return await _fileRepository.DownloadAsync(query);
+                return (stream, fileInfo.Metadata);
             }
             catch (Exception ex)
             {
-                throw new Exception("Error downloading file.", ex);
+                throw new Exception("Error downloading file with metadata.", ex);
             }
         }
 
+
         public async Task<GridFSFileInfo> GetFileInfoByVersionAsync(string name,
-            string owner, string project,
-            long version)
+            string owner, string type,
+            string project, long version)
         {
+            if (version < -1 || version == 0)
+            {
+                throw new ArgumentException("Version cannot be less than -1 or equal to 0.");
+            }
+
             try
             {
+                long versionToUse = version != -1
+                                    ? version
+                                    : await GetLastVersionNumberAsync(name, owner, type, project);
+
                 var query = new BsonDocument
                 {
                     { "filename", name },
                     { "metadata.owner", owner },
-                    { "metadata.project", project }
+                    { "metadata.file_type", type },
+                    { "metadata.project", project },
+                    { "metadata.version_number", versionToUse }
                 };
-
-                if (version != -1)
-                {
-                    query.Set("metadata.version_number", version);
-                }
 
                 return await _fileRepository.GetOneAsync(query);
             }
@@ -103,7 +135,8 @@ namespace AmazingFileVersionControl.Core.Services
         }
 
         public async Task<List<GridFSFileInfo>> GetFileInfoAsync(string name,
-            string owner, string project)
+            string owner, string type,
+            string project)
         {
             try
             {
@@ -111,6 +144,7 @@ namespace AmazingFileVersionControl.Core.Services
                 {
                     { "filename", name },
                     { "metadata.owner", owner },
+                    { "metadata.file_type", type },
                     { "metadata.project", project }
                 };
 
@@ -122,7 +156,25 @@ namespace AmazingFileVersionControl.Core.Services
             }
         }
 
-        public async Task<List<GridFSFileInfo>> GetAllOwnerFilesInfoAsync(string owner)
+        public async Task<List<GridFSFileInfo>> GetProjectFilesInfoAsync(string owner, string project)
+        {
+            try
+            {
+                var query = new BsonDocument
+                {
+                    { "metadata.owner", owner },
+                    { "metadata.project", project}
+                };
+
+                return await _fileRepository.GetManyAsync(query);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Error getting all files info by owner.", ex);
+            }
+        }
+
+        public async Task<List<GridFSFileInfo>> GetAllFilesInfoAsync(string owner)
         {
             try
             {
@@ -140,17 +192,28 @@ namespace AmazingFileVersionControl.Core.Services
         }
 
         public async Task UpdateFileInfoByVersionAsync(string name,
-            string owner, string project,
-            long version, BsonDocument updatedMetadata)
+            string owner, string type,
+            string project, long version,
+            BsonDocument updatedMetadata)
         {
+            if (version < -1 || version == 0)
+            {
+                throw new ArgumentException("Version cannot be less than -1 or equal to 0.");
+            }
+
             try
             {
+                long versionToUse = version != -1
+                                    ? version
+                                    : await GetLastVersionNumberAsync(name, owner, type, project);
+
                 var query = new BsonDocument
                 {
                     { "filename", name },
                     { "metadata.owner", owner },
+                    { "metadata.file_type", type },
                     { "metadata.project", project},
-                    { "metadata.version_number", version}
+                    { "metadata.version_number", versionToUse }
                 };
 
                 await _fileRepository.UpdateManyAsync(query, updatedMetadata);
@@ -162,14 +225,34 @@ namespace AmazingFileVersionControl.Core.Services
         }
 
         public async Task UpdateFileInfoAsync(string name,
-            string owner, string project,
-            BsonDocument updatedMetadata)
+            string owner, string type,
+            string project, BsonDocument updatedMetadata)
         {
             try
             {
                 var query = new BsonDocument
                 {
                     { "filename", name },
+                    { "metadata.owner", owner },
+                    { "metadata.file_type", type },
+                    { "metadata.project", project}
+                };
+
+                await _fileRepository.UpdateManyAsync(query, updatedMetadata);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Error updating file info.", ex);
+            }
+        }
+
+        public async Task UpdateFileInfoByProjectAsync(string owner,
+           string project, BsonDocument updatedMetadata)
+        {
+            try
+            {
+                var query = new BsonDocument
+                {
                     { "metadata.owner", owner },
                     { "metadata.project", project}
                 };
@@ -182,7 +265,7 @@ namespace AmazingFileVersionControl.Core.Services
             }
         }
 
-        public async Task UpdateAllOwnerFilesInfoAsync(string owner,
+        public async Task UpdateAllFilesInfoAsync(string owner,
             BsonDocument updatedMetadata)
         {
             try
@@ -201,17 +284,27 @@ namespace AmazingFileVersionControl.Core.Services
         }
 
         public async Task DeleteFileByVersionAsync(string name,
-            string owner, string project,
-            long version)
+            string owner, string type,
+            string project, long version)
         {
+            if (version < -1 || version == 0)
+            {
+                throw new ArgumentException("Version cannot be less than -1 or equal to 0.");
+            }
+
             try
             {
+                long versionToUse = version != -1
+                                    ? version
+                                    : await GetLastVersionNumberAsync(name, owner, type, project);
+
                 var query = new BsonDocument
                 {
                     { "filename", name },
                     { "metadata.owner", owner },
+                    { "metadata.file_type", type },
                     { "metadata.project", project },
-                    { "metadata.version_number", version }
+                    { "metadata.version_number", versionToUse }
                 };
 
                 await _fileRepository.DeleteOneAsync(query);
@@ -223,7 +316,8 @@ namespace AmazingFileVersionControl.Core.Services
         }
 
         public async Task DeleteFileAsync(string name,
-            string owner, string project)
+            string owner, string type,
+            string project)
         {
             try
             {
@@ -231,6 +325,7 @@ namespace AmazingFileVersionControl.Core.Services
                 {
                     { "filename", name },
                     { "metadata.owner", owner },
+                    { "metadata.file_type", type },
                     { "metadata.project", project }
                 };
 
@@ -242,7 +337,24 @@ namespace AmazingFileVersionControl.Core.Services
             }
         }
 
-        public async Task DeleteAllOwnerFilesAsync(string owner)
+        public async Task DeleteProjectFilesAsync(string owner, string project)
+        {
+            try
+            {
+                var query = new BsonDocument
+                {
+                    { "metadata.owner", owner},
+                    { "metadata.project", project}
+                };
+                await _fileRepository.DeleteManyAsync(query);
+            }
+            catch (Exception ex)
+            {
+                throw new Exception("Error when deleting a project by owner.", ex);
+            }
+        }
+
+        public async Task DeleteAllFilesAsync(string owner)
         {
             try
             {
@@ -259,7 +371,7 @@ namespace AmazingFileVersionControl.Core.Services
         }
 
         private async Task<long> GetLastVersionNumberAsync(string fileName,
-            string owner, string project)
+            string owner, string type, string project)
         {
             try
             {
@@ -267,6 +379,7 @@ namespace AmazingFileVersionControl.Core.Services
                 {
                     { "filename", fileName },
                     { "metadata.owner", owner },
+                    { "metadata.file_type", type },
                     { "metadata.project", project }
                 };
                 var files = await _fileRepository.GetManyAsync(query);
@@ -279,7 +392,7 @@ namespace AmazingFileVersionControl.Core.Services
             }
             catch (Exception ex)
             {
-                throw new Exception("Error getting next version number.", ex);
+                throw new Exception("Error getting last version number.", ex);
             }
         }
     }
